git: Switch merge.conflictStyle to diff3
diff3 seems to be a better alternative to the default merge style. If
this causes issues we should revert.

I am making this switch because I saw the recommendation for diff3 from
the git-delta tool.

Note that I previously tried the new zdiff3 style and had some issues
with massive conflict markers. I will try diff3 first to see if I like
it.