nixos: streamfox-s: Update streamfox-live
Update streamfox-live to the latest version.

In this version we can now specify a single multiplexed WebRTC port,
therefore remove the old port range mapping and multiplex all WebRTC
communication on port 9002.

In addition, we can now specify a HTTP port, so specify 9001 to place
this next to the WebRTC port, and change the nginx proxy config to point
to this port.