nixos: zsh: Add zsh-vi-mode
Add zsh-vi-mode plugin for vi-style editing on the command line.

Include a custom startup script to switch the keybindings to my
preferred jkl; layout instead of hjkl.

Unfortunately, I also used my own fork of zsh-vi-mode to remove the
history keybindings, as this kept overriding fzf's keybindings and I
couldn't find a way to restore the fzf behaviour otherwise - if I find a
way to do so, I can revert back to the original project then.